Background process to automatically create new issues
Every 12 hours checks if a new issue has to be created for a newsletter.
For this it fetches the latest issue of each newsletter and then
compares the maximum sponsor date and the due date of the next issue,
if due date is before maximum it creates the new issue.